refactor(occurrence): rename  `occurrence` folder
