+++ src/components/occurrence-overview/Occurrence.module.scss
