+++ src/components/occurrence-overview/prices/OccurrencePrices.module.scss
