+++ src/components/occurrence-overview/prices/OccurrencePrices.tsx
