+++ src/components/occurrence-overview/rating/OccurrenceRating.module.scss
