+++ src/components/occurrence-overview/rating/OccurrenceRating.tsx
