+++ src/components/occurrence-overview/rating/index.ts
